Fixes a malformedURI error
My webpack CSS-loader started throwing an error today that `vue-file-agent.css` has a malformed URI. I tracked it down to this one line here. I don't know why this started today, I've had zero issues with this for the past several months, I can only assume that css-loader changed something today. Anyway, I tested this locally all I did was encode the svg uri and it works again.

Honestly I'm not sure if this is a *me* issue or not, but I figure if the encoding fixes an edge case and it still works for everyone else than it might be good to let you see it and decide if you want to pull in these changes. 

Thanks for this product, it's awesome!